Update and rename file3_non_differentiable_function.py to file3_Non_Differentiable_Function.py
